Update RunningOutOfEnergy.js
Updated checkEnergy function so that it calls restart when energy drops below zero and regular play is specified.  The if statement in checkEnergy is not being ran when expected, will need to look into it.
Also added a function to update the energy field based of off the distance moved.