Created check for artifacts and win game functions
Created a function called checForArtifact that checks the player's position to see if they are in the space of an artifact.  There are different cases for each artifact but they all return true,  if there is no artifact this function will return false.  Also created the win game function which is the case for the Koka Kola artifact, it displays a message to the user and restarts the game.